FIX moved test after refactoring
